#day15
Question: The head of a linked list and an integer val, remove all the nodes of the linked list that has Node.val == val, and return the new head.

Example 1:
Input: head = [1,2,6,3,4,5,6], val = 6
Output: [1,2,3,4,5]

Example 2:
Input: head = [], val = 1
Output: []
 
Testcase:
Your input: [1,1,7,6,7,9,6]
6
Output: [1,1,7,7,9]
Expected: [1,1,7,7,9]

Constraints:
The number of nodes in the list is in the range [0, 104].
1 <= Node.val <= 50
0 <= val <= 50

Description:
We can use while loop for checking every element on the linked. Inside while loop we will use if condition to check our the element present at that position is equal to the value input. We can use a remove() Method. It is used to remove an element from a linked list. The element is removed from the beginning or head of the linked list.