#day36
Problem: Two arrays arr1 and arr2, the elements of arr2 are distinct, and all elements in arr2 are also in arr1 Sort the elements of arr1 such that the relative ordering of items in arr1 are the same as in arr2. Elements that do not appear in arr2 should be placed at the end of arr1 in ascending order.
Example 1:
Input: arr1 = [2,3,1,3,2,4,6,7,9,2,19], arr2 = [2,1,4,3,9,6]
Output: [2,2,2,1,4,3,3,9,6,7,19]
Example 2:
Input: arr1 = [28,6,22,8,44,17], arr2 = [22,28,8,6]
Output: [22,28,8,6,17,44]

Testcase:
Your input
[2,3,1,3,2,4,6,7,9,2,19]
[2,1,4,3,9,6]
[28,6,22,8,44,17]
[22,28,8,6]
Output
[2,2,2,1,4,3,3,9,6,7,19]
[22,28,8,6,17,44]
Expected
[2,2,2,1,4,3,3,9,6,7,19]
[22,28,8,6,17,44]

Explanation:
Create a temporary array temp of size m and copy the contents of A1[] to it.
Create another array visited[] and initialize all entries in it as false. visited[] is used to mark those elements in temp[] which are copied to A1[].
Sort temp[]
Initialize the output index ind as 0.
Do following for every element of A2[i] in A2[] 
Binary search for all occurrences of A2[i] in temp[], if present then copy all occurrences to A1[ind] and increment ind. Also mark the copied elements visited[]
Copy all unvisited elements from temp[] to A1[]