#day38
Problem: Given a string s and an integer k. You can choose one of the first k letters of s and append it at the end of the string. Return the lexicographically smallest string you could have after applying the mentioned step any number of moves.

Example 1:
Input: s = "cba", k = 1
Output: "acb"
Explanation: 
In the first move, we move the 1st character 'c' to the end, obtaining the string "bac".
In the second move, we move the 1st character 'b' to the end, obtaining the final result "acb".

Example 2:
Input: s = "baaca", k = 3
Output: "aaabc"
Explanation: 
In the first move, we move the 1st character 'b' to the end, obtaining the string "aacab".
In the second move, we move the 3rd character 'c' to the end, obtaining the final result "aaabc".

Testcase:
Your input
"cba"
1
"baaca"
3
Output
"acb"
"aaabc"
Expected
"acb"
"aaabc"

Explanation:
If K = 1, only rotations of S are possible, and the answer is the smallest rotation.
If K > 1, any permutation of S is possible, and the answer is the letters of S written in lexicographic order.