#day39
Problem: A string s, sort it in decreasing order based on the frequency of the characters. The frequency of a character is the number of times it appears in the string.
Return the sorted string. If there are multiple answers, return any of them. 

Example 1:
Input: s = "tree"
Output: "eert"
Explanation: 'e' appears twice while 'r' and 't' both appear once.
So 'e' must appear before both 'r' and 't'. Therefore "eetr" is also a valid answer.

Example 2:
Input: s = "cccaaa"
Output: "aaaccc"
Explanation: Both 'c' and 'a' appear three times, so both "cccaaa" and "aaaccc" are valid answers.
Note that "cacaca" is incorrect, as the same characters must be together.

Testcase:
Your input
"tree"
Output
"eert"
Expected
"eert"

Explanation:
The easiest approach is to use HashMap to solve this problem. First, we traverse a string and put each character and it's count in a HashMap. After that sort the HashMap by values. The time complexity of this approach is O(nlogn) and it's space complexity is O(n).