#day49
 Problem: An array of characters letters that is sorted in non-decreasing order, and a character target. There are at least two different characters in letters. Return the smallest character in letters that is lexicographically greater than target. If such a character does not exist, return the first character in letters.

Example 1:
Input: letters = ["c","f","j"], target = "a"
Output: "c"
Explanation: The smallest character that is lexicographically greater than 'a' in letters is 'c'.

Example 2:
Input: letters = ["c","f","j"], target = "c"
Output: "f"
Explanation: The smallest character that is lexicographically greater than 'c' in letters is 'f'.

Testcase:
Your input
["c","f","j"]
"a"
["c","f","j"]
"c"
["x","x","y","y"]
"z"
Output
"c"
"f"
"x"
Expected
"c"
"f"

Explanation:
We have a list of sorted characters' letters. This is containing only lowercase letters, now we have a target letter t, we have to find the smallest element in the list that is larger than the given target. And letters also wrap around. So, if the target is t = 'z' and letters = ['a', 'b'], the answer is 'a'.