#day51
Problem: An array nums of size n, return the majority element. The majority element is the element that appears more than ⌊n / 2⌋ times. You may assume that the majority element always exists in the array. 

Example 1:
Input: nums = [3,2,3]
Output: 3

Example 2:
Input: nums = [2,2,1,1,1,2,2]
Output: 2

Testcase:
Your input
[3,2,3]
[2,2,1,1,1,2,2]
Output
3
2
Expected
3
2

Explanation:
The majority element is an element in an array that occurs more than (size/2) times in an array (where size​ is the number of elements stored in an array).
Rather than counting occurrences for all the values, let's just count occurrences for the majority elements in each half of the list. And as a bonus: if each half has the same majority element, then that's our majority element for the whole list.