#day54
Problem: We define a harmonious array as an array where the difference between its maximum value and its minimum value is exactly 1. Given an integer array nums, return the length of its longest harmonious subsequence among all its possible subsequences. A subsequence of array is a sequence that can be derived from the array by deleting some or no elements without changing the order of the remaining elements. 

Example 1:
Input: nums = [1,3,2,2,5,2,3,7]
Output: 5
Explanation: The longest harmonious subsequence is [3,2,2,2,3].

Example 2:
Input: nums = [1,2,3,4]
Output: 2

Testcase:
Your input
[1,3,2,2,5,2,3,7]
[1,2,3,4]
[1,1,1,1]
Output
5
2
0
Expected
5
2

Explanation:
As we know a harmonious sequence array is an array where the difference between its maximum value and its minimum value is exactly 1. So, if the input is like [1,3,2,2,5,2,3,7], then the output will be 5, as the longest harmonious subsequence is [4,3,3,3,4].