#day60
Problem: The root of a binary tree, check whether it is a mirror of itself (i.e., symmetric around its center).

Example 1:
Input: root = [1,2,2,3,4,4,3]
Output: true

Example 2:
Input: root = [1,2,2,null,3,null,3]
Output: false

Testcase:
Your input
[1,2,2,3,4,4,3]
[1,2,2,null,3,null,3]
Output
true
false
Expected
true
false

Explanation:
A symmetric binary tree is a tree which is the mirror image of itself, which means we have to check whether the left and right nodes of the tree are the same or not. A Boolean function will initially check for the left node and the right node. If the nodes are empty or NULL, then it will return True.