#day79
Problem: The root of a binary tree, return its maximum depth. A binary tree's maximum depth is the number of nodes along the longest path from the root node down to the farthest leaf node.

Example:
Input: root = [3,9,20,null,null,15,7]
Output: 3

Testcase:
Accepted
Input
root =
[1,null,2]
Output
2
Expected
2

Explanation:
In other words, it is the height of a binary tree. The maximum depth, or height, of this tree is 4; node 7 and node 8 are both four nodes away from the root.