#day84
Problem: A return s[i] is a lower-case letter and s[i + 1] is the same letter but in upper-case or vice-versa.To make the string good, you can choose two adjacent characters that make the string bad and remove them. You can keep doing this until the string becomes good. Return the string after making it good. The answer is guaranteed to be unique under the given constraints. Notice that an empty string is also good. 

Example 1:
Input: s = "leEeetcode"
Output: "leetcode"
Explanation: In the first step, either you choose i = 1 or i = 2, both will result "leEeetcode" to be reduced to "leetcode".

Example 2:
Input: s = "abBAcC"
Output: ""
Explanation: We have many possible scenarios, and all lead to the same answer. For example:
"abBAcC" --> "aAcC" --> "cC" --> ""
"abBAcC" --> "abBA" --> "aA" --> ""

Testcase:
Accepted
Input
s =
"s"
Output
"s"
Expected
"s"

Explanation:
We can iterate from the first character of the given string and append the character to our result string untill it is not bad.
Before appending the current character we would check if appending this character to res string would make it bad or not by comparing it with the last character of res string. If integral difference (ascii) between those two characters is equal to 32 then it is bad combination otherwise it is good.