#day86
Problem: An n x n binary matrix image, flip the image horizontally, then invert it, and return the resulting image. To flip an image horizontally means that each row of the image is reversed. For example, flipping [1,1,0] horizontally results in [0,1,1].
To invert an image means that each 0 is replaced by 1, and each 1 is replaced by 0.
For example, inverting [0,1,1] results in [1,0,0].
 
Example:
Input: image = [[1,1,0],[1,0,1],[0,0,0]]
Output: [[1,0,0],[0,1,0],[1,1,1]]
Explanation: First reverse each row: [[0,1,1],[1,0,1],[0,0,0]].
Then, invert the image: [[1,0,0],[0,1,0],[1,1,1]]

Testcase:
Accepted
Input
image =
[[1,1,0,0],[1,0,0,1],[0,1,1,1],[1,0,1,0]]
Output
[[1,1,0,0],[0,1,1,0],[0,0,0,1],[1,0,1,0]]
Expected
[[1,1,0,0],[0,1,1,0],[0,0,0,1],[1,0,1,0]]

Explanation:
A flipped image or reversed image, the more formal term, is a static or moving image that is generated by a mirror-reversal of an original across a horizontal axis (a flopped image is mirrored across the vertical axis).