#day87
Problem:  the root of a binary tree, return the sum of all left leaves. A leaf is a node with no children. A left leaf is a leaf that is the left child of another node.

Example 1:
Input: root = [3,9,20,null,null,15,7]
Output: 24
Explanation: There are two left leaves in the binary tree, with values 9 and 15 respectively.

Testcase:
Accepted
Input
root =
[1]
Output
0
Expected
0

Explanation:
We will start from the root and check if the node has the left child, if yes then we will check if the node is a leaf node, if it is a leaf node then we will add it to the answer. We will push all the children of the current node in the queue and iteratively visit all the children.