#day88
Problem:  an array nums of non-negative integers. nums is considered special if there exists a number x such that there are exactly x numbers in nums that are greater than or equal to x. Notice that x does not have to be an element in nums. Return x if the array is special, otherwise, return -1. It can be proven that if nums is special, the value for x is unique.

Example:
Input: nums = [3,5]
Output: 2
Explanation: There are 2 values (3 and 5) that are greater than or equal to 2.

Testcase:
Accepted
Input
nums =
[0,0]
Output
-1
Expected
-1

Explanation:
The array in ascending order. For every element arr[i], using binary search get the count of all the elements that are greater than or equal to arr[i] except arr[i] itself. If the count is equal to arr[i] then increment the result. The value of the result in the end.