#day89
Problem: An m x n matrix, return true if the matrix is Toeplitz. Otherwise, return false. A matrix is Toeplitz if every diagonal from top-left to bottom-right has the same elements.

Example:
Input: matrix = [[1,2,3,4],[5,1,2,3],[9,5,1,2]]
Output: true
Explanation:
In the above grid, the diagonals are:
"[9]", "[5, 5]", "[1, 1, 1]", "[2, 2, 2]", "[3, 3]", "[4]".
In each diagonal all elements are the same, so the answer is True.

Testcase:
Accepted
Input
matrix =
[[1,2],[2,2]]
Output
false
Expected
false

Explanation:
A finite Toeplitz matrix is an n × n matrix with the following structure: [1] The entries depend on the difference i − j and hence they are constant down all the diagonals. There are two cases when the determinant is easy to compute. One is when the matrix is upper- or lower-triangular and the determinant is a 0 n .