#day97
Problem: The root of a binary search tree and an integer k, return true if there exist two elements in the BST such that their sum is equal to k, or false otherwise.

Example:
Input: root = [5,3,6,2,4,null,7], k = 9
Output: true

Testcase:
Accepted
Input
root =
[5,3,6,2,4,null,7]
k =
28
Output
false
Expected
false

Explanation:
Suppose we have a Binary Search Tree and one target value; we have to check whether there exist two elements in the BST such that their sum is equal to the given target or not. So, if the input is like. then the output will be True.